feat(zombies): POST /kill + PATCH endpoints, remove DELETE-as-kill
Spec §6 (F4 grounding decision): replace the legacy DELETE-as-kill verb
with a clean POST /v1/workspaces/{ws}/zombies/{id}/kill, plus add
PATCH /v1/workspaces/{ws}/zombies/{id} for config updates. Both publish
control-stream signals so the watcher reacts in real time.

NEW handlers:
  src/http/handlers/zombies/kill.zig (103L)
    UPDATE core.zombies SET status='killed' + control_stream.publish(
    .zombie_status_changed{status=killed}). Watcher flips per-zombie
    cancel flag → executor.cancelExecution(if mid-tool-call) → thread
    exits ≤200ms after the 202. Best-effort publish: row update is
    source of truth, publish failure logs warn but does not roll back.

  src/http/handlers/zombies/patch.zig (125L)
    Partial update of zombie config_json. UPDATE core.zombies +
    control_stream.publish(.zombie_config_changed{config_revision}).
    config_revision = updated_at (BIGINT, ms epoch) — strictly monotonic
    per zombie, no schema migration. Empty body is a 200 no-op.

REMOVED: src/http/handlers/zombies/delete.zig — DELETE-as-kill was a
legacy verb shortcut from before M40. Pre-v2.0 era (VERSION 0.29.0):
no 410 stub per RULE EP4; the old path 404s cleanly.

Routing changes (kept in lockstep across the 3 sources of truth):
  src/http/router.zig
    - delete_workspace_zombie: REMOVED
    - patch_workspace_zombie:  NEW (matches /v1/.../zombies/{id} no suffix)
    - kill_workspace_zombie:   NEW (matches /v1/.../zombies/{id}/kill)
  src/http/route_table.zig + route_table_invoke.zig
    - invokeDeleteWorkspaceZombie: REMOVED
    - invokePatchWorkspaceZombie:  NEW (PATCH only)
    - invokeKillWorkspaceZombie:   NEW (POST only)
  src/http/route_manifest.zig
    - DELETE /v1/.../zombies/{id}: REMOVED
    - PATCH  /v1/.../zombies/{id}: ADDED
    - POST   /v1/.../zombies/{id}/kill: ADDED
  src/http/router_test.zig: rewritten DELETE test → PATCH + /kill assertions

OpenAPI (kept in parity with the manifest, lint-enforced):
  public/openapi/paths/zombies.yaml
    - delete: ... at /v1/.../zombies/{id}: REPLACED with patch:
    - new path /v1/.../zombies/{id}/kill: with post:
  public/openapi/root.yaml
    - new entry $ref-ing the /kill path

zombiectl:
  zombiectl/src/lib/api-paths.js:    +wsZombieKillPath helper
  zombiectl/src/commands/zombie.js:  commandKill now POST + /kill,
                                     requires <zombie_id> (was optional).
  zombiectl/test/zombie.unit.test.js: rewrote DELETE test to assert
                                     POST + /kill suffix; added
                                     missing-arg test.

control_stream.zig: config_revision u32 → i64 (so updated_at ms epoch
fits without truncation). Adjusts decoder, struct field, test, and
revision_buf sizing. control_stream_test.zig updates the type assertion.

ZIG_RULES audit (per the new workflow that landed alongside this PR):
  - Atomic ordering         N/A (no atomics in §6 diff).
  - Mutex pairing           N/A (no mutexes).
  - pub-symbol audit        ✓ inner{Kill,Patch}Zombie used via api.zig
                             facade by route_table_invoke.
  - Allocator-ownership     N/A (no struct owns heap memory).
  - /// doc-comments        ✓ added on innerKillZombie + innerPatchZombie.
  - Milestone-ID gate       ✓ clean.
  - Length gate             ✓ kill 103L / patch 125L / api facade 27L.

Verification:
  zig build test --summary all  ✓ 12/12; 1363 passed; 142 skipped; 0 failed.
  make lint                      ✓ all gates green (router/manifest/openapi
                                  parity now sees POST /kill + PATCH; old
                                  DELETE path no longer in any of the three).
  make test                      ✓ All unit tests passed (Zig + zombiectl
                                  + website + app); zombiectl kill test
                                  rewritten for new POST + /kill semantics.

+++ src/zombie/control_stream_test.zig
@@ -101,7 +101,7 @@ test "decodeEntry: zombie_config_changed parses revision" {
     switch (decoded.message) {
         .zombie_config_changed => |m| {
             try std.testing.expectEqualStrings("z-cfg", m.zombie_id);
-            try std.testing.expectEqual(@as(u32, 42), m.config_revision);
+            try std.testing.expectEqual(@as(i64, 42), m.config_revision);
         },
         else => return error.WrongVariant,
     }
